requirements.txt
Flask
newspaper3k
nltk
numpy
matplotlib
transformers
google-api-python-client
scikit-learn
